+++ tests/conftest.py
@@ -2,10 +2,35 @@ import json
 
 import pytest
 from pydantic import ValidationError
+from rdflib import FOAF
+from yarl import URL
 
+import yaml_ld
 from ldtest.models import TestCase
 from tests.errors import FailureToFail
 from yaml_ld.errors import YAMLLDError
+
+
+URLS = (
+    URL('https://prefix.cc/context.jsonld'),
+    URL('https://schema.org'),
+    URL('https://dbpedia.org/data/Arthur_C._Clarke.json'),
+    URL('https://dbpedia.org/data/Arthur_C._Clarke.jsonld'),
+    URL('https://dbpedia.org/resource/Arthur_C._Clarke'),
+    URL('https://www.wikidata.org/wiki/Special:EntityData/Q42.jsonld'),
+    URL('https://id.loc.gov/authorities/names/n79081644.jsonld'),
+    URL(str(FOAF)),
+)
+
+
+def _url_to_id(url: URL) -> str:
+    return str(url.with_scheme('')).lstrip('/')
+
+
+@pytest.fixture(params=URLS, ids=_url_to_id)
+def url(request) -> URL:
+    yaml_ld.load_document(request.param)
+    return request.param
 
 
 @pytest.fixture()
