+++ yaml_ld/errors.py
@@ -1,6 +1,5 @@
 import textwrap
 from dataclasses import dataclass
-from pathlib import Path
 from typing import Iterable
 
 import funcy
